flask/requirements.txt
Flask==1.0.2
flask-cors
flask-admin==1.5.1
flask-login==0.4.1
flask-mail==0.9.1
blinker==1.4
wtforms==2.2.1
sqlalchemy==1.2.8
gunicorn==19.7.1
marshmallow==3.0.0b7
flake8==3.5.0
arrow==0.10.0
celery==4.2.0
redis==2.10.6
billiard==3.5.0.3
kombu==4.2.1
amqp==2.3.2
vine==1.1.4
pymysql==0.9.2
mysql-connector==2.1.6